Improve certificates for lower tier sole traders and partnerships
https://eaflood.atlassian.net/browse/RUBY-834

The CertificatePresenter tries to make more use of the "main people" names when the business type is sole trader or partnership. However, for lower tier regs, we never ask for the main people. This meant the certificate sometimes displayed blank values where a person's name would be.

So this change only applies that extra logic for upper tier partnerships and sole traders.