Cancel an edit
https://eaflood.atlassian.net/browse/RUBY-832

It should be possible to cancel an in-progress edit.

Cancelling an edit destroys the EditRegistration without making any changes to the original Registration.

If a user returns to edit after cancelling, it creates a fresh EditRegistration with none of the previously cancelled changes.

We ask the user for confirmation before cancelling.