Copy contact name to contact address when completing edit
Spotted this had been left out of the edit completion service. The contact name should be included in the contact address for postal reasons.

If the contact name is edited, or the contact address is edited, this would previously become out of sync. So we add a method to copy over the name, as we do in the RenewalCompletionService.

In the future we should refactor so we don't store this data in two places!