Add support for feature toggles in the service
There are times when a feature is not complete or is not ready for general use but we still want to ship.

We have found adding support for feature toggles to our services solves the problem.

The first step is adding a general `FeatureToggle` class to the engine which handles reading optional features from a YAML file, and whether they are enabled or not.

Adding it to the engine means both the front and back office apps can then use it to determine if a feature should be made available.

N.B. This is essentially a straight copy & paste from [Waste Exemptions PR #222](https://github.com/DEFRA/waste-exemptions-engine/pull/222)