Store a past_registration when completing an edit
https://eaflood.atlassian.net/browse/RUBY-831

When a user makes an edit to a registration, we should store a new past_registration object containing the pre-edit version of the registration.

This will let us keep track of changes made by the edit.

This should happen at the point of completion. Cancelling an edit does not save a past_registration.