Add field to store cause of past registration
If a past registration is created due to an edit, we store this in a new `cause` field.

We also use this to make sure we always save past regs when they're because of an edit, as previously this could be blocked if there was already an existing one with the same expiry date.