Build new registration workflow for Upper tier
From: https://eaflood.atlassian.net/browse/RUBY-915

This PR includes the code necessary to create an upper tier workflow for new registrations.
The majority of the work is on the test suite, while the rest of the work is on the state machine and the used forms are all already been built and shared with Edit and Renewals.