Fix matching comp. reg. no's with lowercase chars
https://eaflood.atlassian.net/browse/RUBY-942

We had a report from NCCC of a renewal that got stuck at the Companies house number page. Even though the registration number is the same as the one on the original registration, the renewal journey kept directing the user to the 'This registration cannot be renewed' page.

Upon investigation it turned out to be because the number in the registration had lowercase characters, and the number in the renewal had uppercase. We were able to reproduce the issue using the following steps.

- `registration` is completed using old service and has a registration number of `ni123456`
- user starts renewal. `ni123456` is copied to the `transient_registration`
- user gets to the company registration number screen and sees `ni123456`
- user clicks submit but gets a *Companies House couldn't find a company with this number validation error*
- user corrects the number to be `NI123456`. This is valid and recognised by the lookup so the value is persisted to the `transient_registration`
- our workflow engine runs its `require_new_registration_based_on_company_no?` test
- it compares `ni123456` to `NI123456` and returns `true`
- user directed to the *You cannot renew page*

The companies house check won’t accept the registration number being lowercase. And the check to see if the number has changed isn’t ignoring case so thinks it has.

Hence this change. Here we are just updating the underpinning logic in `require_new_registration_based_on_company_no` to handle differences in case.