Remove business owner screen from Lower Tier journey
From: https://eaflood.atlassian.net/browse/RUBY-836

We discovered today that the wireframe was wrong in the fact that we should capture the business owner details for lower tier registrations. After having a conversation with Anrew and Emi we agreed that the old system is correct and that the wireframe needed fixing so that we don't ask about Business owner details on the lower tier registration journey.