Add presenter for order details
https://eaflood.atlassian.net/browse/RUBY-833

This is part of a set of multiple PRs to configure payment when a carrier type changes during editing.

This presenter will be used to display the individual charges of an order, as well as the total cost.

The aim is to standardise our various payment pages as much as we can.