Move checks on registration type changes to concern
https://eaflood.atlassian.net/browse/RUBY-833

This is part of a set of multiple PRs to configure payment when a carrier type changes during editing.

Renewals already had a check to see if a registration type had been changed from the original registration. We'll need to use this for edits as well, so it makes sense to move this check to a new shared concern.