Make EditCompletionService deal with new finance info
https://eaflood.atlassian.net/browse/RUBY-833

This is part of a set of multiple PRs to configure payment when a carrier type changes during editing.

The EditCompletionService currently disregards any finance_details on the transient object. However, if the user is required to pay for an edit, a new finance_details will be created. We need to be able to merge any new orders and payments onto the original registration.

The OrderCopyCardsCompletionService already deals with this, so we just need to make some tweaks the EditCompletionService to follow that example.