Build finance_details object for edit journey
https://eaflood.atlassian.net/browse/RUBY-833

This is part of a set of multiple PRs to configure payment when a carrier type changes during editing.

This PR sets up a BuildEditFinanceDetailsService class which is used to create finance_details objects for the edit journey. It's pretty similar to OrderAdditionalCardsService so there's probably room for consolidating once we have a consistent way of building this objects (I'm looking at you, renewals).