Set up workflow for edit payments by bank transfer
https://eaflood.atlassian.net/browse/RUBY-833

This is part of a set of multiple PRs to configure payment when a carrier type changes during editing.

This PR sets up the workflow and pages for paying for an edit by bank transfer. This includes the payment summary and bank transfer info pages.

WorldPay will be added separately in a future PR.