Add fetch_presenters hook to FormsController
https://eaflood.atlassian.net/browse/RUBY-833

Added to enable us to properly use presenters in the edit payment workflow.

This will let us override this method as needed in subclassed forms so we can create presenters at the right time.

Previously we just stuck presenters in the `new` action of the individual controllers. However, if the form failed to submit (for example, a validation error occurred) then the engine would try to `render :new` as part of the `create` action. But in this case the presenter would not be created and so the page would not render properly.

The downside of this is that the `new` method in FormsController no longer returns the result of `set_up_form` - instead it returned the content of `fetch_presenters`, which is `nil` by default. So any actions in subclassed controllers which relied on lines like `return unless super(FormClass, "form_class")` returning true were no longer being called, as the `unless` no longer passed. Setting `new` to return `true` by default fixes this.

However longterm the FormsController just needs some proper refactoring as it's a big ol mess.