Add renewal number lookup page
https://eaflood.atlassian.net/browse/RUBY-841

When external users enter the service, they see a page that asks what the user wants to do - start a new registration or renew. If they select 'renew', they see this page: 'What is your waste carrier registration number?'

We need to replace this page from the frontend with a new version.