Update text on 'application received' pages
https://eaflood.atlassian.net/browse/RUBY-958

This makes a minor text change to the 'renewal received' and 'registration received' pages that are seen when a user submits an application that is pending a conviction check.

Previously this included a mention of payment which could confuse users who had already paid. So this PR updates the text to clarify it.