Send email when new reg application is pending conviction check
https://eaflood.atlassian.net/browse/RUBY-842

If a new registration has been paid for but cannot be completed immediately because of convictions checks, we should still send the user a confirmation email to let them know the application was received.

This email is not sent if there is a pending payment at the end of the new registration journey. In that case we send the pending payment email only.