Move and rename renewing registration workflow specs
This is just a change of directory name from "workflow_states" to the more specific "renewing_registration_workflow", and then taking "transient_registration_" off the start of all the spec files. This brings the specs more into line with the other workflow specs, which didn't exist when they were first created.

There's a lot more we can do to improve these specs, but I'm isolating renaming in this PR to preserve the history as much as possible.