Amend Travis rubocop command
There isn't a problem with the json formatter. It was that we are getting a rubocop error we couldn't see initially locally or in the build.

After some research found that we can tell rubocop to use multiple formatters in one command. So now updating the Travis build to output stuff to file for use by sonarcloud, and stuff to STDOUT so we can see any issues in the build.