Merge repeating request specs
We have a lot of request specs which are very time-expensive to run. It now takes over 4 minutes to run the unit test suite locally.

To speed up the test suite, any request tests which make the same request repeatedly (eg submitting a form with the same parameters multiple times to test individual things) should be merged together.