Add bank transfer details to new reg completion screen
https://eaflood.atlassian.net/browse/RUBY-1005

Users expect the bank transfer details to be on the final confirmation screen once they have submitted their registration. This change adds the details to the confirmation page for new registrations. Renewals will be updated in a separate PR.

This works in combination with https://github.com/DEFRA/waste-carriers-engine/pull/820, which removes the bank transfer info from the preceding page in both journeys.