Implement renew via magic link journey happy path
From:  https://eaflood.atlassian.net/browse/RUBY-987

This implements the redirect from the renew controller for magic link to the renewal journey.
It also address issues with login, since we do not want a user to login if they have started from the magic link, but we do want them to login if they then pick the journey app again from the normal flow. A flag on the renewing registration has been added to achieve that.