Add DataLayerHelper for tracking different journeys
https://eaflood.atlassian.net/browse/RUBY-950

We want to be able to track the performance of the new registration journey and the renewals journey. Because they use the same pages, we'll need a way to differentiate other than just using the URLs.

This PR will add a `data_layer` method that we will use in the front office header. This will be used by Google Tag Manager to send pageviews with the journey as a custom dimension to Google Analytics.

The initial plan was to put this helper in the front office as we do with WEX - however due to changes in the engines' ApplicationControllers, we've decided to put it in the engine instead.