Add tiers to data layer
https://eaflood.atlassian.net/browse/RUBY-857

We were previously able to separate page views for new registrations and renewals in Google Analytics based on their different URL patterns. However, when the new registration journey is moved into the engine, many pages will have the same URLs. But we want to still be able to separate the pageviews for these journeys.

We can do this by adding a custom dimension to the data layer called "Tier" and configuring Google Tag Manager to pass that value to Google Analytics with the pageview. This allows us to filter pageviews by user journey when needed.

Since some early pages in the new registration journey won't yet have a tier, we use `unknown` as the default.