Pin mongoid-locker to 2.0.0
The front-office and back-office apps when the engine has been bumped have also taken on the latest version of mongoid-locker (v2.0.1) because it was not pinned.

Though only a minor change it appears it has a dependency on Mongoid v7 and changes it has made around not using ActiveSupport method `delegate` since version 7.1

We ascertained this from a statement made in the issue [mongoid 7.1.0 issue with arguments](https://github.com/mongoid/mongoid-locker/issues/85#issuecomment-609033576)

> Mongoid gem doesn't use ActiveSupport method delegate since 7.1 version.

The issue was fixed in the mongoid-locker PR [Add forwardable module and upgrade rubocop](https://github.com/mongoid/mongoid-locker/pull/86). And that change seems to be the key difference between version 2.0.0 and 2.0.1.

The error we see in the apps when v2.0.1 is being used is

```
An error occurred while loading ./spec/forms/waste_carriers_engine/base_forms_spec.rb.
Failure/Error: delegate :contact_address, :contact_email, :registered_address, to: :registration
ArgumentError:
  wrong number of arguments (given 4, expected 1)
```

In this example if you go to order_copy_cards_registration.rb:10 you'll find

```ruby
    delegate :contact_address, :contact_email, :registered_address, to: :registration
```

Depending on whether you run rspec or try to hit a page in the app this is the error that gets thrown.

So we are pinning the version of mongoid-locker we use, until we can complete the work to upgrade the version of ruby and rails we are using. We should then be able to use Mongoid v7.1 (or greater) and hopefully the issue goes away!