Remove link to confirm email address
https://eaflood.atlassian.net/browse/RUBY-1131

When a user goes to the front office signin screen and clicks "Resend a confirmation email" they get redirected to the frontend.

Since we are getting rid of the frontend and new registrations in the front office won't require an account, we have decided to get rid of this.

Why this code is in the engine is another question...