Promote little FeatureToggle to be a proper Model
From: https://eaflood.atlassian.net/browse/RUBY-1118

Given that little FeatureToggle has behaved very well in the past few months, we decided to upgrade him to be a proper Model. XD

This model will be used by the new defra-ruby-features engine. Toggles will be persisted in the database.

Things to notice is, in order to handle the fact that the model won't have persisted data about existing toggles, previously persisted in a yml file for read-only, the code will handle that case by redirecting the query for a toggle status to the content of the yml file.
By doing this we will just need to deploy and then take our time creating the toggles through the back office interface. In the meantime, everything will work just fine.

There will need some cleanup to be done at the subsequent deploy, but this still looks like a cleaner solution than a rake task, so that we don't need to cleanup places like Jenkins.