Move feature toggle config to engine
https://eaflood.atlassian.net/browse/RUBY-1118

We have created a number of gems to support reuse of functionality across our services. If the functionality is to be used across both host apps, or because of other functionality, we'll incorporate the gem into the engine. [defra-ruby-alert](https://github.com/DEFRA/defra-ruby-alert) is an example of this.

Our typical pattern when we do this is to then control any config for that gem from the engine, and if extra info is needed from the host apps to pass it via the engines config.

[defra-ruby-features](https://github.com/DEFRA/defra-ruby-features) has been implemented by referencing and configuring the gem directly in the host apps. So to make things consistent this change updates the implementation to match our existing pattern of controlling everything from the engine.