Don't copy expires_on around when editing
https://eaflood.atlassian.net/browse/RUBY-1151

Previously during an edit, the value of expires_on was copied from the registration to the transient_registration, and then copied back again when the edit was completed.

If the registration was renewed between the edit being started and completed, this resulted in a bug where the expires_on date was "reset" to its previous value.

This PR stops us from copying the value of expires_on into or out of edit_registrations.