Update defra-ruby-alert
DEFRA/defra-ruby-alert#11

The defra-ruby-alert gem was updated to use the latest version of Airbrake at the same time as this project was updated to work with Rails 6 and Ruby 2.7.

What we did not realise is that the Airbrake gem had introduced a performance monitoring feature which is enabled by default. As we log our exceptions to errbit this feature was not supported so our logs are getting spammed with 404 messages from the Airbrake gem.

We've fixed that in defra-ruby-alert and this updates the project to use the latest version.