Remove all reference to receipt email form
There is an argument for keeping this code here. We do intend to come back and implement automatically generating and sending a receipt to everyone, no matter how they pay or what they are paying for.

But with the current phase of WCR coming to a close that work will remain on the backlog for an unknown amount of time. It is very likely if it does get revisted, it will be with a different team who may wish to go in a different direction.

So rather than retaining and maintaining code that is not being used, we're deleting it. We can always revisits the original PR and this one to rebuild the form and page if needed.