Add payment email section to payment summary page
https://eaflood.atlassian.net/browse/RUBY-1141
https://eaflood.atlassian.net/browse/RUBY-1137

Having removed our first design for solving the problem of capturing a user's preferred Worldpay payment confirmation email address (see [PR #889](https://github.com/DEFRA/waste-carriers-engine/pull/889) this implements our second version!

This change adds a hidden section in the payment summary screen which will appear when a user selects 'Pay by credit or debit card'. Along with a header and some content it will display a field, pre-populated with the contact email address. Whatever is in there when submitted will be saved as the `receipt_email`.

This PR will focus on adding the section and validating it correctly. For example, we only care about the field being populated correctly if the user is paying by card. If they are not then we shouldn't be forcing them to correct any invalid changes they have made to the field.

A subsequent PR will focus on the logic of hiding and showing the section at the appropriate times.